refactor: create class for Shake detection to unpollute main
